requirements.txt
scrapy
scrapy-splash
beautifulsoup4
python-dotenv